#117 Started splitting up the Ability tests and did some minor reworking of Pokemon tests to not have single monolithic test cases
